Make Saver aware of power-off events.
This handles the case where the TV is turned off while still on a given
input, and that input goes inactive, then wakes up again.

I'm not 100% certain we _should_ be handling these cases -- after all,
we weren't actually the ones to enable the screensaver? -- but it does
avoid it constantly trying to change to the sasver on a turned-off TV.